requirements.txt
Flask==3.0.3
flask-cors==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.1
google-generativeai==0.7.0
gunicorn==21.2.0
dnspython==2.6.1
